Refactor: 관리자용 API 분리
